test+chore: 11 nouveaux tests + cleanup MTP (review pre-merge)
Tests ajoutes pour combler les gaps de couverture identifies en revue:

1. MaskedEmbedderTests (4 tests)
   - Init shapes (centroids Linear + token_ordering Int32 buffer)
   - Forward shape [B, L, vocab] sans NaN
   - Sparse scatter: positions non-selectionnees == min(selected_logits)-1
   - Batch + multi-position preserves shapes

2. ForwardWithIntermediatesTests (5 tests)
   - TextModel: forwardCollectingIntermediates expose hidden + preNormHidden +
     intermediates avec shapes K/V correctes par layer type (full=512, sliding=256)
   - Parite: callAsFunction == forwardCollectingIntermediates.hidden (delegate)
   - pre-norm != post-norm (norm a fait quelque chose)
   - LanguageModel: forwardWithIntermediates expose logits + hiddenStates +
     preNormHiddenStates + intermediates
   - Parite: callAsFunction logits == forwardWithIntermediates.logits

3. DrafterTrainingTests (2 tests)
   - drafterLoss retourne scalaire fini non-NaN, ntoks = B*(L-2)
   - trainDrafter sur 2 iter ne crash pas (smoke)

Cleanup code review:
- Gemma4MTPPipeline.swift: supprime un bloc `MainActor.run { }` vide (dead code)
- Gemma4MTPPipeline.swift: fix typo "Sharedeading" -> "Lecture"
- Gemma4AssistantDraftModel.swift: supprime un cast Int32 redondant
  (MLXArray([Int32(lastBonus)]) -> MLXArray([lastBonus]) puisque lastBonus
  est deja Int32)

Total: 107 tests existants + 11 nouveaux = 118 tests dans la suite.
Verifies individuellement (Metal Toolchain locale en panne temporaire
empeche le full run mais chaque file passe en isolation).

+++ Sources/Gemma4Swift/Speculative/Gemma4AssistantDraftModel.swift
@@ -288,7 +288,7 @@ public class Gemma4AssistantDraftModel: Module {
         }
         precondition(blockSize >= 2, "blockSize doit etre >= 2 (sinon aucun draft)")
 
-        var tok = MLXArray([Int32(lastBonus)]).reshaped(1, 1)  // [B=1, 1]
+        var tok = MLXArray([lastBonus]).reshaped(1, 1)  // [B=1, 1]
         var hPrev = hidden  // [B, 1, backbone]
         var tokens: [MLXArray] = []
